Move DSP Manager into settings (1/2)
I noticed that most people that used this ROM chose to setup DSP
Manager once and hide it from their app drawer or ignore it. I
think this is because it is more like the Theme Choser and the new
CM Updater and used more as a utility that you dont need to use
often, for this reason I purpose we move it into settings like
the other applications.

Change-Id: Ib0c2da56659768c605c7ec9ffbd966679ad0ca3f